move functions to service-wide utils
