+++ services/backend/arm_backend/main.py
@@ -40,9 +40,9 @@ from arm_backend.routers import (
     transcoder,
     transcodes,
 )
-from arm_backend.routers.system import default_roots, ensure_roots
 from arm_backend.seeders import CONFIG_SINGLETON_ID, run_seeders
 from arm_backend.transcode_dispatcher import TranscodeDispatcher
+from arm_backend.utils import ensure_roots, default_roots
 from arm_backend.ws import WSHub
 from arm_backend.ws.router import router as ws_router
 from arm_common import Config, Gpu, GpuStatus, configure_service_logging
